% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam2.B.3.R
\docType{data}
\name{DataExam2.B.3}
\alias{DataExam2.B.3}
\title{Data for Example 2.B.3 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup (p-55)}
\format{A \code{data.frame} with 6 rows and 2 variables.}
\usage{
data(DataExam2.B.3)
}
\description{
Exam2.B.3 is used to illustrate one way treatment design with Gaussian observations.
}
\details{
\itemize{
       \item trt treatments as factor with number 1 to 3
       \item y   response variable
       }
}
\examples{
data(DataExam2.B.3)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam2.B.3}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\keyword{datasets}
