% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam2.B.4.R
\docType{data}
\name{DataExam2.B.4}
\alias{DataExam2.B.4}
\title{Data for Example 2.B.4 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup (p-54)}
\format{A \code{data.frame} with 6 rows and 4 variables.}
\usage{
data(DataExam2.B.4)
}
\description{
Exam2.B.4 is used to illustrate one way treatment design with Binomial observations.
}
\details{
\itemize{
       \item obs number of observations
       \item trt three treatments with class factor
       \item Nij  number of bernouli trials on each individual
       \item y  number of successes on each individual
       }
}
\examples{
data(DataExam2.B.4)
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam2.B.4}}
}
\keyword{datasets}

