% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam2.B.7.R
\docType{data}
\name{DataExam2.B.7}
\alias{DataExam2.B.7}
\title{Data for Example 2.B.7 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup (p-60)}
\format{A \code{data.frame} with 16 rows and 4 variables.}
\usage{
data(DataExam2.B.7)
}
\description{
Exam2.B.7 is related to multi batch regression data assuming different forms of linear models with factorial experiment.
}
\details{
\itemize{
       \item Rep number of replications
       \item a factor with two levels 1 and 2
       \item b factor with two levels 1 and 2
       \item y  response variable
       }
}
\examples{
data(DataExam2.B.7)
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam2.B.7}}
}
\keyword{datasets}

