% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet3.1.R
\docType{data}
\name{DataSet3.1}
\alias{DataSet3.1}
\title{Data for Example 3.1 and Example 3.2 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\format{A \code{data.frame} with 20 rows and 5 variables.}
\usage{
data(DataSet3.1)
}
\description{
DataSet3.1 is used for linear and generalized linear models
}
\details{
\itemize{
       \item trt two treatment 0 and 1
       \item rep unit of observation or observation ID
       \item Y  is continuous & may be assumed Gaussian
       \item N  is the number of obs
       \item F  is the number of "successes"(N and F specify a binomial response)
       }
}
\examples{
data(DataSet3.1)
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam3.2}}
}
\keyword{datasets}

