% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet3.2.R
\docType{data}
\name{DataSet3.2}
\alias{DataSet3.2}
\title{DataSt3.2 for Example 3.3, Example 3.4, Example3.6, Example3.8 and Example 3.9 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\format{A \code{data.frame} with 32 rows and 10 variables.}
\usage{
data(DataSet3.2)
}
\description{
DataSet3.2 Multi-Location, 4 Treatment Randomized Block
}
\details{
\itemize{
       \item trt              two treatment 0 and 1
       \item loc              four locations used as blocks
       \item Y                is Gaussian response variable
       \item Nbin             subjects at each Loc x Trt for binomial response
       \item S1 and S2        are two binomial response variables
       \item count1 and count 2 used later
       \item A and B          are factors with level 0 and 1
            }
}
\examples{
data(DataSet3.2)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam3.3}}
   \code{\link{Exam3.9}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\keyword{datasets}
