% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet3.3.R
\docType{data}
\name{DataSet3.3}
\alias{DataSet3.3}
\title{Data for Example3.7 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\format{A \code{data.frame} with 36 rows and 6 variables.}
\usage{
data(DataSet3.3)
}
\description{
Exam1.2 is used to see types of model effects by plotting regression data
}
\details{
\itemize{
       \item X     Each batch observed at several times:0,3,6,12,24,36,48 months
       \item Y     continuous variable observed at each level of X
       \item Fav   number of successes
       \item N     isndependent bernoulli trials
       \item Batch Batches as 1, 2, 3, 4
       \item Count binomial response variable
       }
}
\examples{
data(DataSet3.3)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC press.
 }
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\keyword{datasets}
