% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet4.1.R
\docType{data}
\name{DataSet4.1}
\alias{DataSet4.1}
\title{Data for Example 4.1 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\format{A \code{data.frame} with 60 rows and 3 variables.}
\usage{
data(DataSet4.1)
}
\description{
DataSet4.1 comes from Cochran and Cox (1957) Experimental Design
}
\details{
\itemize{
       \item blocks 15 blocks in an incomplete block desgin
       \item trt    treatments representing incomplete block desgin
       \item y      is continuous & may be assumed Gaussian
       }
}
\examples{
data(DataSet4.1)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC press.
\item Cochran, W. G., & Cox, G. M. (1957). \emph{Experimental designs}.
 }
}
\seealso{
\code{\link{Exam4.1}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\keyword{datasets}
