% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet5.2.R
\docType{data}
\name{DataSet5.2}
\alias{DataSet5.2}
\title{Data for Example 5.2 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\format{A \code{data.frame} with 9 rows and 4 variables.}
\usage{
data(DataSet5.2)
}
\description{
DataSet5.2 is used for three factor orthogonal main effects only design with sequential fitting of predictors
}
\details{
\itemize{
       \item a  is predictor variable  with level 0, 1
       \item b  is predictor variable  with level 0, 1
       \item c  is predictor variable  with level 0, 1
       \item y  response variable
       }
}
\examples{
data(DataSet5.2)
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam5.2}}
}
\keyword{datasets}

