% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet7.4.R
\docType{data}
\name{DataSet7.4}
\alias{DataSet7.4}
\title{Data for Example 7.4 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\usage{
data(DataSet7.4)
}
\description{
Data for Example 7.4 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup
}
\examples{
data(DataSet7.4)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC press.
 }
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\keyword{datasets}
