% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet8.1.R
\docType{data}
\name{DataSet8.1}
\alias{DataSet8.1}
\title{Data for Example 8.1 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\format{
A \code{data.frame} with 30 rows and 4 variables.
}
\usage{
data(DataSet8.1)
}
\description{
DataSet8.1 is used for Nested factorial structure
}
\details{
\itemize{
       \item block  10 blocks
       \item trt    6 treatments nested within sets
       \item set    2 sets
       \item y      is a Gaussian response variable
       }
}
\examples{
data(DataSet8.1)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear Mixed Models: Modern Concepts, Methods and Applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam8.1}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
        Adeela Munawar (\email{adeela.uaf@gmail.com})
}
\keyword{datasets}
