% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet8.2.R
\docType{data}
\name{DataSet8.2}
\alias{DataSet8.2}
\title{Data for Example 8.2 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\format{
A \code{data.frame} with 36 rows and 6 variables.
}
\usage{
data(DataSet8.2)
}
\description{
DataSet8.2 is used for Incomplete strip-plot ( 3 cross 3 factorial).
}
\details{
\itemize{
       \item block  9 blocks each consisting of 2 rows and 2 coloumns
       \item a      is a factor with 3 levels assigned at random to rows
       \item b      is a factor with 3 levels assigned at random to columns
       \item y      is a Gaussian response variable
       }
}
\examples{
data(DataSet8.2)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear Mixed Models: Modern Concepts, Methods and Applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam8.2}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
        Adeela Munawar (\email{adeela.uaf@gmail.com})
}
\keyword{datasets}
