% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet8.3.R
\docType{data}
\name{DataSet8.3}
\alias{DataSet8.3}
\title{Data for Example 8.3 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\format{
A \code{data.frame} with 28 rows and 4 variables.
}
\usage{
data(DataSet8.3)
}
\description{
DataSet8.3 is used for Response surface design with incomplete blocking
}
\details{
\itemize{
       \item block  with 7 blocks
       \item a      is a factor with 3 levels 0,-1 and 1
       \item b      is a factor with 3 levels 0,-1 and 1
       \item c      is a factor with 3 levels 0,-1 and 1
       \item y      is a Gaussian response variable
       }
}
\examples{
data(DataSet8.3)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear Mixed Models: Modern Concepts, Methods and Applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam8.3}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
        Adeela Munawar (\email{adeela.uaf@gmail.com})
}
\keyword{datasets}
