% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet9.1.R
\docType{data}
\name{DataSet9.1}
\alias{DataSet9.1}
\title{Data for Example 9.1 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\format{
A \code{data.frame} with 24 rows and 2 variables.
}
\usage{
data(DataSet9.1)
}
\description{
DataSet9.1 is used for One-way random effects only model
}
\details{
\itemize{
       \item a  is a factor with 12 levels
       \item y  is a Gaussian response variable
       }
}
\examples{
data(DataSet9.1)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear Mixed Models: Modern Concepts, Methods and Applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam9.1}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
        Adeela Munawar (\email{adeela.uaf@gmail.com})
}
\keyword{datasets}
