% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet9.2.R
\docType{data}
\name{DataSet9.2}
\alias{DataSet9.2}
\title{Data for Example 9.2 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\format{
A \code{data.frame} with 28 rows and 3 variables with levels of b nested within levels of.
}
\usage{
data(DataSet9.2)
}
\description{
DataSet9.2 is used for Two way random effects nested model
}
\details{
\itemize{
       \item a  is a factor with 7 levels
       \item b  is a factor with 2 levels
       \item y  is a Gaussian response variable
       }
}
\examples{
data(DataSet9.2)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear Mixed Models: Modern Concepts, Methods and Applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam9.2}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
        Adeela Munawar (\email{adeela.uaf@gmail.com})
}
\keyword{datasets}
