% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet9.4.R
\docType{data}
\name{DataSet9.4}
\alias{DataSet9.4}
\title{Data for Example 9.4 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\format{
A \code{data.frame} with 32 rows and 3 variables
}
\usage{
data(DataSet9.4)
}
\description{
DataSet9.4 is used for Relationship between BLUP and Fixed Effect Estimators
}
\details{
\itemize{
       \item a  is a factor with 2 levels
       \item b  is a factor with 8 levels
       \item y  is a Gaussian response variable
       }
}
\examples{
data(DataSet9.4)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear Mixed Models: Modern Concepts, Methods and Applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam9.4}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
        Adeela Munawar (\email{adeela.uaf@gmail.com})
}
\keyword{datasets}
