% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exam2.B.6.R
\docType{data}
\name{Exam2.B.6}
\alias{Exam2.B.6}
\title{Example 2.B.6 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup(p-58)}
\description{
Exam2.B.6 is related to multi batch regression data assuming different forms of linear models keeping batch effect random.
}
\examples{
#-----------------------------------------------------------------------------------
## Nested Model with no intercept
#-----------------------------------------------------------------------------------
data(Table1.2)
library(nlme)
Table1.2$Batch <- factor(x = Table1.2$Batch)
Exam2.B.6fm1 <-
  lme(
      fixed       = Y~X
    , data        = Table1.2
    , random      = list(Batch = pdDiag(~1), X = pdDiag(~1))
    , correlation = NULL
    , weights     = NULL
  # , subset
    , method      = "REML" #c("REML", "ML")
    , na.action   = na.fail
  # , control     = list()
    , contrasts   = NULL
    , keep.data   = TRUE
  )
}
\references{
\enumerate{
\item Stroup, W. W. (2012).
     \emph{Generalized Linear Mixed Models: Modern Concepts, Methods and Applications}.
       CRC Press.
 }
}
\seealso{
\code{\link{Table1.2}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\keyword{datasets}
