% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exam7.2.R
\docType{data}
\name{Exam7.2}
\alias{Exam7.2}
\title{Example 7.2 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup (p-219)}
\description{
Exam7.2 explains multifactor models with some factors qualitative and some quantitative(Equal slopes ANCOVA)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).
     \emph{Generalized Linear Mixed Models: Modern Concepts, Methods and Applications}.
       CRC Press.
 }
}
\seealso{
\code{\link{DataSet7.2}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\keyword{datasets}
