% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GriffingData2.R
\docType{data}
\name{GriffingData2}
\alias{GriffingData2}
\title{Data for Diallel Analysis using Griffing Approach Method 2}
\format{A \code{data.frame} with 144 rows and 4 variables.}
\usage{
data(GriffingData2)
}
\description{
\code{Griffing} is used for performing Diallel Analysis using Griffing's Approach.
}
\details{
\itemize{
       \item Cross1 Cross 1
       \item Cross2 Cross 2
       \item Rep Replicate
       \item Yield Yield Response
       }
}
\examples{
data(GriffingData2)
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\references{
\enumerate{
\item Griffing, B. (1956) Concept of General and Specific Combining Ability
            in relation to Diallel Crossing Systems. \emph{Australian Journal of Biological Sciences},
             \strong{9(4)}, 463--493.
\item Singh, R. K. and Chaudhary, B. D. (2004) \emph{Biometrical Methods in Quantitative Genetic Analysis}.
             New Delhi: Kalyani.
 }
}
\seealso{
\code{\link{Griffing}}
 , \code{\link{GriffingData1}}
 , \code{\link{GriffingData3}}
 , \code{\link{GriffingData4}}
}
\keyword{datasets}

