% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaymanData.R
\docType{data}
\name{HaymanData}
\alias{HaymanData}
\title{Data for Diallel Analysis using Hayman's Approach}
\format{A \code{data.frame} with 256 rows and 4 variables.}
\usage{
data(HaymanData)
}
\description{
\code{Griffing} is used for performing Diallel Analysis using Hayman's Approach.
}
\details{
\itemize{
       \item Cross1 Cross 1
       \item Cross2 Cross 2
       \item Rep Replicate
       \item Yield Yield Response
       }
}
\examples{
data(HaymanData)
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
}
\references{
\enumerate{
\item Griffing, B. (1956) Concept of General and Specific Combining Ability
            in relation to Diallel Crossing Systems. \emph{Australian Journal of Biological Sciences},
             \strong{9(4)}, 463--493.
 \item Test
 }
}
\keyword{datasets}

