% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.1.R
\docType{data}
\name{Table1.1}
\alias{Table1.1}
\title{Data for Table1.1 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup}
\format{A \code{data.frame} with 11 rows and 3 variables.}
\usage{
data(Table1.1)
}
\description{
\code{Exam1.1} is used for inspecting probability distribution and to define a plausible process.
}
\details{
\itemize{
       \item x independent variable
       \item Nx bernouli trials(bernouli outcomes on each individual)
       \item y  number of successes on each individual
       }
}
\examples{
data(Table1.1)
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam1.1}}
}
\keyword{datasets}

