% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.2.R
\docType{data}
\name{Table1.2}
\alias{Table1.2}
\title{Data for Table1.2 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup(p-10)}
\format{
A \code{data.frame} with 36 rows and 5 variables.
}
\usage{
data(Table1.2)
}
\description{
Exam1.2 is used to see types of model effects by plotting regression data
}
\details{
\itemize{
       \item X     have 11 levels in varying intervals from 0 to 48 observed for multiple batches
       \item Y     continuous variable observed at each level of X
       \item Fav   number of successes
       \item N     number of bernoulli trials
       \item Batch Batches as 1, 2, 3, 4
       }
}
\examples{
data(Table1.2)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC press.
 }
}
\seealso{
\code{\link{Exam1.2}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\keyword{datasets}
