% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam2.B.2.R
\docType{data}
\name{DataExam2.B.2}
\alias{DataExam2.B.2}
\title{Data for Example 2.B.2 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup (p-54)}
\format{A \code{data.frame} with 11 rows and 3 variables.}
\usage{
data(DataExam2.B.2)
}
\description{
Exam2.B.2 is used to visualize the effect of glm model statement with binomial data with logit and probit links.
}
\details{
{
       \item{\code{x}}{independent variable}
       \item{\code{n}}{bernouli trials(bernouli outcomes on each individual)}
       \item{\code{y}}{number of successes on each individual}
       }
}
\examples{
data(DataExam2.B.2)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC press.
 }
}
\seealso{
\link{Exam2.B.2}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\keyword{datasets}
