% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exam7.6.2.1.R
\docType{data}
\name{Exam7.6.2.1}
\alias{Exam7.6.2.1}
\title{Example 7.6.2.1 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Walter W. Stroup (p-231)}
\description{
Exam7.6.2.1 Nonlinear Mean Models ( Quantitative by quantitative models)
}
\references{
\enumerate{
\item Stroup, W. W. (2012).
     \emph{Generalized Linear Mixed Models: Modern Concepts, Methods and Applications}.
       CRC Press.
 }
}
\seealso{
\code{\link{DataSet7.6}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\keyword{datasets}
